remove UserManager and Password attribute
