+++ mdw-common/src/com/centurylink/mdw/model/user/User.java
@@ -38,7 +38,6 @@ public class User implements Serializable, Comparable<User>, Jsonable {
     // new friendly attribute names that can be displayed as labels
     public static final String EMAIL_ADDRESS = "Email Address";
     public static final String PHONE_NUMBER = "Phone Number";
-    public static final String PASSWORD = "Password";
     // old unfriendly attribute names for compatibility
     public static final String OLD_EMAIL_ADDRESS = "emailAddress";
     public static final String OLD_PHONE_NUMBER = "phoneNumber";
@@ -318,10 +317,8 @@ public class User implements Serializable, Comparable<User>, Jsonable {
         if (attributes != null) {
             JSONObject attrsJson = create();
             for (String attr : attributes.keySet()) {
-                if (!attr.equals(PASSWORD)) {
-                    String value = attributes.get(attr);
-                    attrsJson.put(attr, value == null ? "" : value);
-                }
+                String value = attributes.get(attr);
+                attrsJson.put(attr, value == null ? "" : value);
             }
             json.put("attributes", attrsJson);
         }
