--- mdw-common/src/com/centurylink/mdw/util/ReleaseHelper.java
@@ -1,8 +1,0 @@
-/**
- * Copyright (c) 2014 CenturyLink, Inc. All Rights Reserved.
- */
-package com.centurylink.mdw.util;
-
-public class ReleaseHelper {
-
-}
