+++ mdw-services/src/com/centurylink/mdw/services/UserServices.java
@@ -15,6 +15,8 @@
  */
 package com.centurylink.mdw.services;
 
+import java.util.List;
+
 import com.centurylink.mdw.dataaccess.DataAccessException;
 import com.centurylink.mdw.model.user.RoleList;
 import com.centurylink.mdw.model.user.UserAction;
@@ -63,4 +65,6 @@ public interface UserServices {
 
     public void auditLog(UserAction userAction) throws DataAccessException;
 
+    public List<User> getWorkgroupUsers(List<String> groups) throws DataAccessException;
+    public List<String> getWorkgroupEmails(List<String> groups) throws DataAccessException;
 }
